+++ ElvUI_MerathilisUI/Core/Changelog/6.76.lua
@@ -1,8 +1,15 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[676] = {
-	RELEASE_DATE = "TBD",
+	RELEASE_DATE = "11.10.2025",
 	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	NEW = {
+		"[Skins]: Added an initial Plumber Skin",
+	},
+	IMPROVEMENTS = {
+		"[AutoButtons]: Added Legion Remix openables",
+		"[Minimap Buttons]: Added a sort feature",
+		"[Skins]: Updated Friends List skin to also support now Recent Allies",
+		"[Skins]: Updated SilverDragon skin to support the mount button",
+	},
 }