+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Changes ###
 
-  * [Core]: Small updates due to 11.2.5 changes
-  * [Achievement Tracker]: Added a new Achievement Tracker module
-  * [Skins]: Finished Rematch skin
+  * [AutoButtons]: Added Legion Remix openables
+  * [Minimap Buttons]: Added a sort feature
+  * [Skins]: Added an initial Plumber Skin
+  * [Skins]: Updated Friends List skin to also support now Recent Allies
+  * [Skins]: Updated SilverDragon skin to support the mount button